requirements.txt
openai
pdfplumber
python-docx
python-dotenv
requests
pytest
reportlab
